refactor: site metadata
